Improve type conversion

+++ src/Gradient.php
@@ -380,7 +380,7 @@ abstract class Gradient extends \Com\Tecnick\Pdf\Graph\Raw
                     $bpcd,
                     ((($point - $coords_min) / ($coords_max - $coords_min)) * $bpcd)
                 ));
-                $this->gradients[$ngr]['stream'] .= chr(floor($point / 256)).chr(floor($point % 256));
+                $this->gradients[$ngr]['stream'] .= chr((int)floor($point / 256)).chr((int)floor($point % 256));
             }
             foreach ($par['colors'] as $color) {
                 //each color component as 8 bit
